refactor: code quality cleanup (#81)
