--- .github/workflows/pre-commit.yml
@@ -1,24 +1,0 @@
-name: pre-commit
-
-on:
-  push:
-    branches: [main]
-  pull_request:
-    branches: [main]
-  workflow_dispatch:
-
-jobs:
-  pre-commit:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-
-      - name: Install Nix
-        uses: DeterminateSystems/nix-installer-action@main
-
-      - name: Set up Nix cache
-        uses: DeterminateSystems/magic-nix-cache-action@main
-
-      - name: Run pre-commit hooks via prek
-        run: nix develop --command prek run --all-files
