--- .github/workflows/tests.yml
@@ -1,119 +1,0 @@
-name: test suite
-run-name: test suite for ${{ github.event.pull_request.title || github.ref }}
-on:
-  workflow_dispatch:
-  push:
-    branches:
-      - main
-  pull_request:
-    types:
-      - opened
-      - reopened
-      - synchronize
-      - ready_for_review
-
-concurrency:
-  group: ${{ github.workflow }}-${{ github.ref }}
-  cancel-in-progress: true
-
-env:
-  DENO_V: 2.3.5
-  GHJK_VERSION: "v0.3.2"
-  GHJK_ENV: "ci"
-
-jobs:
-  changes:
-    runs-on: ubuntu-latest
-    permissions:
-      pull-requests: read
-    steps:
-      - uses: actions/checkout@v6
-      - uses: dorny/paths-filter@v3
-        id: filter
-        with:
-          filters: |
-            full:
-              - '.github/workflows/tests.yml'
-              - 'src/**'
-              - 'tests/**'
-              - 'examples/**'
-    outputs:
-      full: ${{ steps.filter.outputs.full }}
-
-  pre-commit:
-    needs: changes
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v6
-      - uses: actions/setup-python@v6
-        with:
-          python-version: "3.x"
-      - uses: denoland/setup-deno@v2
-        with:
-          deno-version: ${{ env.DENO_V }}
-      - name: Install tofu
-        run: |
-          curl --proto '=https' --tlsv1.2 -fsSL https://get.opentofu.org/install-opentofu.sh -o install-opentofu.sh
-          chmod +x install-opentofu.sh
-          ./install-opentofu.sh --install-method deb
-          rm -f install-opentofu.sh
-
-      - shell: bash
-        run: |
-          python -m pip install --upgrade pip
-          pip install pre-commit
-          pre-commit install
-          deno --version
-          pre-commit run --all-files
-
-  test-full:
-    needs: [changes]
-    if: ${{ needs.changes.outputs.full == 'true' && github.event.pull_request.draft == false }}
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v6
-      - uses: denoland/setup-deno@v2
-        with:
-          deno-version: ${{ env.DENO_V }}
-      - name: Download Install Script
-        run: curl -fsSL "https://raw.github.com/metatypedev/ghjk/$GHJK_VERSION/install.sh" -o install.sh
-      - name: Execute Install Script
-        run: yes | bash install.sh
-      - run: echo "$HOME/.local/bin" >> "$GITHUB_PATH"
-      - run: echo "BASH_ENV=$HOME/.local/share/ghjk/env.sh" >> "$GITHUB_ENV"
-      - uses: actions/setup-python@v6
-        with:
-          python-version: "3.x"
-      - name: Install tofu
-        run: |
-          curl --proto '=https' --tlsv1.2 -fsSL https://get.opentofu.org/install-opentofu.sh -o install-opentofu.sh
-          chmod +x install-opentofu.sh
-          ./install-opentofu.sh --install-method deb
-          rm -f install-opentofu.sh
-      - uses: actions/setup-node@v6
-        with:
-          node-version: 18
-      - name: setup start-server-and-test
-        run: npm install -g start-server-and-test
-      - shell: bash
-        env:
-          AUTH_TYPE: "default"
-          LOG_LEVEL: "DEBUG"
-          ENV: "DEV"
-          S3_ACCESS_KEY: ${{ secrets.S3_ACCESS_KEY }}
-          S3_SECRET_KEY: ${{ secrets.S3_SECRET_KEY }}
-          OPENSTACK_USERNAME: ${{ secrets.OPENSTACK_USERNAME }}
-          OPENSTACK_PASSWORD: ${{ secrets.OPENSTACK_PASSWORD }}
-          OPENSTACK_PROJECT: ${{ secrets.OPENSTACK_PROJECT }}
-          AWS_ACCESS_KEY_ID: ${{ secrets.OPENSTACK_USERNAME }}
-          AWS_SECRET_ACCESS_KEY: ${{ secrets.OPENSTACK_PASSWORD }}
-        run: |
-          # run all tests
-          deno --version
-          ghjk x dev-compose s3
-          sleep 20
-
-          deno install
-
-          # ghjk x setup-auth
-          npx start-server-and-test 'deno serve -A --unstable-kv src/main.ts' http://0.0.0.0:8000/ 'deno test -A'
