+++ chart/values.yaml
@@ -1,4 +1,3 @@
-
 name: herald
 namespace: herald
 
@@ -87,9 +86,9 @@ ingress:
         - path: /
           pathType: ImplementationSpecific
   tls: []
-    # - secretName: web-tls
-    #   hosts:
-    #     - chart-example.local
+# - secretName: web-tls
+#   hosts:
+#     - chart-example.local
 
 volumeMounts:
   - name: herald
