+++ deno.jsonc
@@ -30,6 +30,11 @@
     "cliffy/ansi/": "https://deno.land/x/cliffy@v1.0.0-rc.3/ansi/"
   },
   "compilerOptions": {},
+  "fmt": {
+    "exclude": [
+      "./chart/"
+    ]
+  },
   "lint": {
     "exclude": [
       "x",
