+++ src/Api.ts
@@ -1,8 +1,11 @@
 import { HttpApi, OpenApi } from "@effect/platform";
-import { HealthApi } from "./Frontend/Health/Api.ts";
-import { S3Api } from "./Frontend/Api.ts";
+import { HealthHttpApi } from "./Frontend/Health/Api.ts";
+import { HttpS3Api } from "./Frontend/Api.ts";
 
-export class Api extends HttpApi.make("api")
-  .add(HealthApi)
-  .add(S3Api)
+// the http interface is declared first and separately
+// and the impl is to adhere to it
+// used for openAPI
+export class HttpHeraldApi extends HttpApi.make("HeraldHttpApi")
+  .add(HealthHttpApi)
+  .add(HttpS3Api)
   .annotate(OpenApi.Title, "Herald API") {}
