+++ src/Backends/Swift/Client.ts
@@ -1,6 +1,7 @@
-import { Context, Effect, Layer, type Schema } from "effect";
+import { Cache, Context, Effect, Layer, type Schema } from "effect";
+import { HttpClient, HttpClientRequest } from "@effect/platform";
 import type { MaterializedBucket, SwiftConfig } from "../../Domain/Config.ts";
-import { AppConfig } from "../../Config/Layer.ts";
+import { HeraldConfig } from "../../Config/Layer.ts";
 
 export interface SwiftAuthMeta {
   readonly token: string;
@@ -35,160 +36,152 @@ interface SwiftTokenResponse {
 
 export const SwiftClientLive = Layer.effect(
   SwiftClient,
-  AppConfig.pipe(
-    Effect.flatMap((appConfig) => {
-      const cache = new Map<string, SwiftAuthMeta & { expires: number }>();
+  Effect.gen(function* () {
+    const appConfig = yield* HeraldConfig;
+    const client = yield* HttpClient.HttpClient;
 
-      const fetchAuthMeta = (
-        config: Schema.Schema.Type<typeof SwiftConfig>,
-      ): Effect.Effect<SwiftAuthMeta, Error, never> => {
-        const { auth_url, credentials, region } = config;
+    const fetchAuthMeta = (
+      config: Schema.Schema.Type<typeof SwiftConfig>,
+    ): Effect.Effect<SwiftAuthMeta, Error, never> => {
+      const { auth_url, credentials, region } = config;
 
-        if (!auth_url) {
-          return Effect.fail(
-            new Error("auth_url is required for Swift backend"),
+      if (!credentials || !("username" in credentials)) {
+        return Effect.fail(
+          new Error(
+            "Swift credentials (username, password, etc.) are required",
+          ),
+        );
+      }
+
+      const {
+        username,
+        password,
+        project_name,
+        user_domain_name = "Default",
+        project_domain_name = "Default",
+      } = credentials;
+
+      const requestBody = {
+        auth: {
+          identity: {
+            methods: ["password"],
+            password: {
+              user: {
+                name: username,
+                domain: { name: user_domain_name },
+                password: password,
+              },
+            },
+          },
+          scope: {
+            project: {
+              domain: { name: project_domain_name },
+              name: project_name,
+            },
+          },
+        },
+      };
+
+      return Effect.gen(function* () {
+        const request = yield* HttpClientRequest.post(`${auth_url}/auth/tokens`)
+          .pipe(
+            HttpClientRequest.bodyJson(requestBody),
+            Effect.mapError((e) => new Error(String(e))),
+          );
+        const response = yield* client.execute(request).pipe(
+          Effect.mapError((e) => new Error(String(e))),
+        );
+
+        if (response.status < 200 || response.status >= 300) {
+          const msg = yield* response.text.pipe(
+            Effect.orElseSucceed(() => "Unknown error"),
+          );
+          return yield* Effect.fail(
+            new Error(`Failed to authenticate with Swift: ${msg}`),
           );
         }
-        if (!credentials || !("username" in credentials)) {
-          return Effect.fail(
+
+        const token = response.headers["x-subject-token"];
+        const tokenStr = Array.isArray(token) ? token[0] : token;
+
+        if (!tokenStr) {
+          return yield* Effect.fail(
             new Error(
-              "Swift credentials (username, password, etc.) are required",
+              "X-Subject-Token header missing from Swift response",
             ),
           );
         }
 
-        const {
-          username,
-          password,
-          project_name,
-          user_domain_name = "Default",
-          project_domain_name = "Default",
-        } = credentials;
+        const body = (yield* response.json.pipe(
+          Effect.mapError((e) => new Error(String(e))),
+        )) as SwiftTokenResponse;
 
-        const requestBody = JSON.stringify({
-          auth: {
-            identity: {
-              methods: ["password"],
-              password: {
-                user: {
-                  name: username,
-                  domain: { name: user_domain_name },
-                  password: password,
-                },
-              },
-            },
-            scope: {
-              project: {
-                domain: { name: project_domain_name },
-                name: project_name,
-              },
-            },
-          },
-        });
+        const catalog = body.token.catalog;
+        const storageService = catalog.find((s) => s.type === "object-store");
 
-        return Effect.tryPromise({
-          try: async () => {
-            const response = await fetch(`${auth_url}/auth/tokens`, {
-              method: "POST",
-              headers: { "Content-Type": "application/json" },
-              body: requestBody,
-            });
+        if (!storageService) {
+          return yield* Effect.fail(
+            new Error(
+              "Object Store service not found in Swift catalog",
+            ),
+          );
+        }
 
-            if (!response.ok) {
-              const msg = await response.text();
-              throw new Error(`Failed to authenticate with Swift: ${msg}`);
-            }
+        const endpoint = storageService.endpoints.find(
+          (e) =>
+            (region ? e.region === region : true) &&
+            e.interface === "public",
+        );
 
-            const token = response.headers.get("X-Subject-Token");
-            if (!token) {
-              throw new Error(
-                "X-Subject-Token header missing from Swift response",
-              );
-            }
+        if (!endpoint) {
+          return yield* Effect.fail(
+            new Error(
+              `Public Swift endpoint not found (region: ${region ?? "any"})`,
+            ),
+          );
+        }
 
-            const body = (await response.json()) as SwiftTokenResponse;
-            const catalog = body.token.catalog;
-            const storageService = catalog.find((s) =>
-              s.type === "object-store"
-            );
+        return {
+          token: tokenStr,
+          storageUrl: endpoint.url,
+        };
+      });
+    };
 
-            if (!storageService) {
-              throw new Error(
-                "Object Store service not found in Swift catalog",
-              );
-            }
+    const cache = yield* Cache.make({
+      capacity: 100,
+      timeToLive: "50 minutes", // Swift tokens usually last 1h
+      lookup: (config: Schema.Schema.Type<typeof SwiftConfig>) =>
+        fetchAuthMeta(config),
+    });
 
-            const endpoint = storageService.endpoints.find(
-              (e) =>
-                (region ? e.region === region : true) &&
-                e.interface === "public",
-            );
+    return SwiftClient.of({
+      getAuthMeta: (
+        bucket: MaterializedBucket | { backend_id: string },
+      ) => {
+        let backend_id: string;
+        let config: Schema.Schema.Type<typeof SwiftConfig>;
 
-            if (!endpoint) {
-              throw new Error(
-                `Public Swift endpoint not found (region: ${region ?? "any"})`,
-              );
-            }
+        if ("protocol" in bucket) {
+          backend_id = bucket.backend_id;
+          config = appConfig.raw.backends[backend_id] as Schema.Schema.Type<
+            typeof SwiftConfig
+          >;
+        } else {
+          backend_id = bucket.backend_id;
+          config = appConfig.raw.backends[backend_id] as Schema.Schema.Type<
+            typeof SwiftConfig
+          >;
+        }
 
-            return {
-              token,
-              storageUrl: endpoint.url,
-            };
-          },
-          catch: (e) => e as Error,
-        });
-      };
+        if (!config || config.protocol !== "swift") {
+          return Effect.fail(
+            new Error(`Backend ${backend_id} is not a Swift backend`),
+          );
+        }
 
-      return Effect.succeed(
-        SwiftClient.of({
-          getAuthMeta: (
-            bucket: MaterializedBucket | { backend_id: string },
-          ) => {
-            let backend_id: string;
-            let config: Schema.Schema.Type<typeof SwiftConfig>;
-
-            if ("protocol" in bucket) {
-              backend_id = bucket.backend_id;
-              config = appConfig.raw.backends[backend_id] as Schema.Schema.Type<
-                typeof SwiftConfig
-              >;
-            } else {
-              backend_id = bucket.backend_id;
-              config = appConfig.raw.backends[backend_id] as Schema.Schema.Type<
-                typeof SwiftConfig
-              >;
-            }
-
-            if (!config || config.protocol !== "swift") {
-              return Effect.fail(
-                new Error(`Backend ${backend_id} is not a Swift backend`),
-              );
-            }
-
-            const cacheKey =
-              `${backend_id}:${config.auth_url}:${config.region}`;
-            const cached = cache.get(cacheKey);
-            const now = Date.now();
-
-            if (cached && cached.expires > now) {
-              return Effect.succeed({
-                token: cached.token,
-                storageUrl: cached.storageUrl,
-              });
-            }
-
-            return fetchAuthMeta(config).pipe(
-              Effect.tap((meta) => {
-                // Cache for 50 minutes (Swift tokens usually last 1h)
-                cache.set(cacheKey, {
-                  ...meta,
-                  expires: now + 50 * 60 * 1000,
-                });
-              }),
-            );
-          },
-        }),
-      );
-    }),
-  ),
+        return cache.get(config);
+      },
+    });
+  }),
 );
