+++ src/Frontend/Buckets/Head.ts
@@ -1,12 +1,10 @@
 import { Effect } from "effect";
 import { HttpServerResponse } from "@effect/platform";
-import { resolveBucket } from "../Utils.ts";
+import { RequestContext } from "../Utils.ts";
 
-export const headBucket = (
-  { path: { bucket } }: { path: { bucket: string } },
-) =>
-  resolveBucket(bucket, (backend) =>
-    Effect.gen(function* () {
-      yield* backend.headBucket();
-      return HttpServerResponse.empty({ status: 200 });
-    }));
+export const headBucket = () =>
+  Effect.gen(function* () {
+    const { backend } = yield* RequestContext;
+    yield* backend.headBucket();
+    return HttpServerResponse.empty({ status: 200 });
+  });
