+++ src/Frontend/Health/Api.ts
@@ -1,7 +1,7 @@
 import { HttpApiEndpoint, HttpApiGroup, OpenApi } from "@effect/platform";
 import { Schema } from "effect";
 
-export class HealthApi extends HttpApiGroup.make("health")
+export class HealthHttpApi extends HttpApiGroup.make("health")
   .add(
     HttpApiEndpoint.get("getStatus", "/health")
       .addSuccess(Schema.Struct({ status: Schema.Literal("ok") })),
