+++ src/Frontend/Objects/Delete.ts
@@ -1,27 +1,20 @@
 import { Effect } from "effect";
-import { HttpServerRequest, HttpServerResponse } from "@effect/platform";
-import { extractKey, resolveBucket } from "../Utils.ts";
+import { HttpServerResponse } from "@effect/platform";
+import { RequestContext } from "../Utils.ts";
 
 /**
  * Handler for DeleteObject (DELETE /:bucket/*)
  */
-export const deleteObject = (
-  { path: { bucket } }: { path: { bucket: string } },
-) =>
-  resolveBucket(bucket, (backend) =>
-    Effect.gen(function* () {
-      const request = yield* HttpServerRequest.HttpServerRequest;
-      const key = extractKey(request.url, bucket);
-      const url = new URL(request.url, "http://localhost");
-      const searchParams = url.searchParams;
+export const deleteObject = () =>
+  Effect.gen(function* () {
+    const { backend, key, params } = yield* RequestContext;
 
-      if (searchParams.has("uploadId")) {
-        // Abort Multipart Upload
-        const uploadId = searchParams.get("uploadId")!;
-        yield* backend.abortMultipartUpload(key, uploadId);
-        return HttpServerResponse.empty({ status: 204 });
-      }
-
-      yield* backend.deleteObject(key);
+    if (params.uploadId) {
+      // Abort Multipart Upload
+      yield* backend.abortMultipartUpload(key, params.uploadId);
       return HttpServerResponse.empty({ status: 204 });
-    }));
+    }
+
+    yield* backend.deleteObject(key);
+    return HttpServerResponse.empty({ status: 204 });
+  });
