+++ src/Frontend/Objects/List.ts
@@ -1,61 +1,49 @@
 import { Effect } from "effect";
-import { HttpServerRequest } from "@effect/platform";
-import { resolveBucket } from "../Utils.ts";
+import { RequestContext } from "../Utils.ts";
 import { S3Xml } from "../../Services/S3Xml.ts";
 
 /**
  * Handler for ListObjects (GET /:bucket)
  */
-export const listObjects = (
-  { path: { bucket } }: { path: { bucket: string } },
-) =>
-  resolveBucket(bucket, (backend) =>
-    Effect.gen(function* () {
-      const request = yield* HttpServerRequest.HttpServerRequest;
-      const s3Xml = yield* S3Xml;
-      const url = new URL(request.url, "http://localhost");
-      const searchParams = url.searchParams;
+export const listObjects = () =>
+  Effect.gen(function* () {
+    const { backend, params } = yield* RequestContext;
+    const s3Xml = yield* S3Xml;
 
-      if (searchParams.has("versions")) {
-        const result = yield* backend.listVersions({
-          prefix: searchParams.get("prefix") ?? undefined,
-          delimiter: searchParams.get("delimiter") ?? undefined,
-          keyMarker: searchParams.get("key-marker") ?? undefined,
-          versionIdMarker: searchParams.get("version-id-marker") ?? undefined,
-          maxKeys: searchParams.has("max-keys")
-            ? parseInt(searchParams.get("max-keys")!)
-            : undefined,
-          encodingType: searchParams.get("encoding-type") ?? undefined,
-        });
-        return s3Xml.formatListVersions(result);
-      }
-
-      if (searchParams.has("uploads")) {
-        const result = yield* backend.listMultipartUploads({
-          prefix: searchParams.get("prefix") ?? undefined,
-          delimiter: searchParams.get("delimiter") ?? undefined,
-          keyMarker: searchParams.get("key-marker") ?? undefined,
-          uploadIdMarker: searchParams.get("upload-id-marker") ?? undefined,
-          maxUploads: searchParams.has("max-uploads")
-            ? parseInt(searchParams.get("max-uploads")!)
-            : undefined,
-          encodingType: searchParams.get("encoding-type") ?? undefined,
-        });
-        return s3Xml.formatListMultipartUploads(result);
-      }
-
-      const result = yield* backend.listObjects({
-        prefix: searchParams.get("prefix") ?? undefined,
-        delimiter: searchParams.get("delimiter") ?? undefined,
-        marker: searchParams.get("marker") ?? undefined,
-        maxKeys: searchParams.has("max-keys")
-          ? parseInt(searchParams.get("max-keys")!)
-          : undefined,
-        encodingType: searchParams.get("encoding-type") ?? undefined,
-        continuationToken: searchParams.get("continuation-token") ?? undefined,
-        startAfter: searchParams.get("start-after") ?? undefined,
-        listType: searchParams.get("list-type") === "2" ? 2 : 1,
+    if (params.versions !== undefined) {
+      const result = yield* backend.listVersions({
+        prefix: params.prefix,
+        delimiter: params.delimiter,
+        keyMarker: params["key-marker"],
+        versionIdMarker: params["version-id-marker"],
+        maxKeys: params["max-keys"],
+        encodingType: params["encoding-type"],
       });
+      return s3Xml.formatListVersions(result);
+    }
 
-      return s3Xml.formatListObjects(result);
-    }));
+    if (params.uploads !== undefined) {
+      const result = yield* backend.listMultipartUploads({
+        prefix: params.prefix,
+        delimiter: params.delimiter,
+        keyMarker: params["key-marker"],
+        uploadIdMarker: params["upload-id-marker"],
+        maxUploads: params["max-uploads"],
+        encodingType: params["encoding-type"],
+      });
+      return s3Xml.formatListMultipartUploads(result);
+    }
+
+    const result = yield* backend.listObjects({
+      prefix: params.prefix,
+      delimiter: params.delimiter,
+      marker: params.marker,
+      maxKeys: params["max-keys"],
+      encodingType: params["encoding-type"],
+      continuationToken: params["continuation-token"],
+      startAfter: params["start-after"],
+      listType: params["list-type"] === "2" ? 2 : 1,
+    });
+
+    return s3Xml.formatListObjects(result);
+  });
