+++ tests/health.test.ts
@@ -5,27 +5,29 @@ import {
   HttpApiClient,
   HttpServer,
 } from "@effect/platform";
-import { Api, HttpHealthLive, HttpS3Live } from "../src/Http.ts";
-import { AppConfig } from "../src/Config/Layer.ts";
+import { HeraldHttpApi, HttpHealthLive, HttpS3Live } from "../src/Http.ts";
+import { HeraldConfig } from "../src/Config/Layer.ts";
 import { S3ClientLive } from "../src/Backends/S3/Client.ts";
+import { SwiftClientLive } from "../src/Backends/Swift/Client.ts";
 import { S3XmlLive } from "../src/Services/S3Xml.ts";
 import { BackendResolverLive } from "../src/Services/BackendResolver.ts";
 import { EffectAssert, testEffect } from "./utils.ts";
 
 testEffect("health/getStatus", () =>
   Effect.gen(function* () {
-    const AppConfigLive = Layer.succeed(AppConfig, {
+    const HeraldConfigLive = Layer.succeed(HeraldConfig, {
       raw: { backends: {} },
       lookupBucket: () => Option.none(),
     });
 
-    const ApiWithRequirements = HttpApiBuilder.api(Api).pipe(
+    const ApiWithRequirements = HttpApiBuilder.api(HeraldHttpApi).pipe(
       Layer.provide(HttpHealthLive),
       Layer.provide(HttpS3Live),
-      Layer.provide(S3ClientLive),
       Layer.provide(BackendResolverLive),
+      Layer.provide(S3ClientLive),
+      Layer.provide(SwiftClientLive),
       Layer.provide(S3XmlLive),
-      Layer.provide(AppConfigLive),
+      Layer.provide(HeraldConfigLive),
       Layer.provide(FetchHttpClient.layer),
       Layer.provideMerge(HttpServer.layerContext),
     );
@@ -34,7 +36,7 @@ testEffect("health/getStatus", () =>
     const webHandler = HttpApiBuilder.toWebHandler(ApiWithRequirements);
 
     const clientProgram = Effect.gen(function* () {
-      const client = yield* HttpApiClient.make(Api, {
+      const client = yield* HttpApiClient.make(HeraldHttpApi, {
         baseUrl: "http://localhost",
       });
       return yield* client.health.getStatus();
